+++ CoronerIntegrations/Plugin.cs
@@ -4,19 +4,23 @@ using CoronerIntegrations.Patch.BiodiversityIntegration;
 using CoronerIntegrations.Patch.BiodiversityIntegration.Aloe;
 using CoronerIntegrations.Patch.BiodiversityIntegration.Critters;
 using CoronerIntegrations.Patch.BiodiversityIntegration.Ogopogo;
+using CoronerIntegrations.Patch.BunkbedReviveIntegration;
 using CoronerIntegrations.Patch.ChillaxScrapsIntegration;
 using CoronerIntegrations.Patch.CountryRoadCreatureIntegration;
+using CoronerIntegrations.Patch.HerobrineIntegration;
 using CoronerIntegrations.Patch.JackensteinApparatusIntegration;
 using CoronerIntegrations.Patch.LegendWeathersIntegration;
 using CoronerIntegrations.Patch.LethalAnomaliesIntegration;
 using CoronerIntegrations.Patch.LethalDoorsFixedIntegration;
 using CoronerIntegrations.Patch.LockerIntegration;
 using CoronerIntegrations.Patch.PremiumScrapsIntegration;
+using CoronerIntegrations.Patch.ReviveCompanyIntegration;
 using CoronerIntegrations.Patch.RollingGiantIntegration;
 using CoronerIntegrations.Patch.ScopophobiaIntegration;
 using CoronerIntegrations.Patch.ShockwaveDroneIntegration;
 using CoronerIntegrations.Patch.SirenHeadIntegration;
 using CoronerIntegrations.Patch.TheCabinetIntegration;
+using CoronerIntegrations.Patch.TheRollingChairIntegration;
 using CoronerIntegrations.Patch.UsualScrapIntegration;
 using CoronerIntegrations.Patch.ZeldaScrapsIntegration;
 using HarmonyLib;
@@ -34,14 +38,19 @@ namespace CoronerIntegrations
     [BepInPlugin(PluginInfo.PLUGIN_GUID, PluginInfo.PLUGIN_NAME, PluginInfo.PLUGIN_VERSION)]
     [BepInDependency("com.elitemastereric.coroner", BepInDependency.DependencyFlags.HardDependency)]
     [BepInDependency("com.github.biodiversitylc.Biodiversity", BepInDependency.DependencyFlags.SoftDependency)]
+    [BepInDependency("viviko.BunkbedRevive", BepInDependency.DependencyFlags.SoftDependency)]
+    [BepInDependency("Erksmit.LethalRevives", BepInDependency.DependencyFlags.SoftDependency)]
+    [BepInDependency("OpJosMod.ReviveCompany", BepInDependency.DependencyFlags.SoftDependency)]
     [BepInDependency("wexop.country_road_creature", BepInDependency.DependencyFlags.SoftDependency)]
     [BepInDependency("com.zealsprince.locker", BepInDependency.DependencyFlags.SoftDependency)]
     [BepInDependency("Scopophobia", BepInDependency.DependencyFlags.SoftDependency)]
+    [BepInDependency("ccode.chair", BepInDependency.DependencyFlags.SoftDependency)]
     [BepInDependency("Ccode.SirenHead", BepInDependency.DependencyFlags.SoftDependency)]
     [BepInDependency("VectorV.TheCabinet", BepInDependency.DependencyFlags.SoftDependency)]
     [BepInDependency("Emil.UsualScrap", BepInDependency.DependencyFlags.SoftDependency)]
     [BepInDependency("nomnomab.rollinggiant", BepInDependency.DependencyFlags.SoftDependency)]
     [BepInDependency("droneenemy", BepInDependency.DependencyFlags.SoftDependency)]
+    [BepInDependency("Kittenji.HerobrineMod", BepInDependency.DependencyFlags.SoftDependency)]
     [BepInDependency("Entity378.LethalDoorsFixed", BepInDependency.DependencyFlags.SoftDependency)]
     [BepInDependency("zigzag.legendweathers", BepInDependency.DependencyFlags.SoftDependency)]
     [BepInDependency("zigzag.premiumscraps", BepInDependency.DependencyFlags.SoftDependency)]
@@ -79,6 +88,21 @@ namespace CoronerIntegrations
 
                 BiodiversitySoftDep.CoronerRegister();
             }
+            
+            PluginLogger.LogInfo($"BunkbedRevive_LegitsFork Found: {BunkbedReviveSoftDep.enabled}");
+            if (BunkbedReviveSoftDep.enabled)
+            {
+                Harmony.PatchAll(typeof(BunkbedRevivePlayerPatch));
+                //No register because there is no cause of death
+            }
+            
+            PluginLogger.LogInfo($"ReviveCompany or Patched Found: {ReviveCompanySoftDep.enabled}"); //Should work for both Erk's and OpJos'
+            if (ReviveCompanySoftDep.enabled)
+            {
+                Harmony.PatchAll(typeof(ReviveCompanyPatch));
+                //No register because there is no cause of death
+            }
+
 
             PluginLogger.LogInfo($"Scopophobia Found: {ScopophobiaSoftDep.enabled}");
             if (ScopophobiaSoftDep.enabled)
@@ -92,6 +116,13 @@ namespace CoronerIntegrations
             {
                 Harmony.PatchAll(typeof(SirenHeadEatPlayerPatch));
                 SirenHeadSoftDep.CoronerRegister();
+            }
+            
+            PluginLogger.LogInfo($"The Rolling Chair Found: {TheRollingChairSoftDep.enabled}");
+            if (TheRollingChairSoftDep.enabled)
+            {
+                Harmony.PatchAll(typeof(RollingChairCollidePatch));
+                TheRollingChairSoftDep.CoronerRegister();
             }
 
             PluginLogger.LogInfo($"The Cabinet Found: {TheCabinetSoftDep.enabled}");
@@ -137,6 +168,15 @@ namespace CoronerIntegrations
                 Harmony.PatchAll(typeof(ShockScanPatch));
                 Harmony.PatchAll(typeof(ShockCreateExplosionPatch));
                 ShockwaveDroneSoftDep.CoronerRegister();
+            }
+            
+            PluginLogger.LogInfo($"HerobrineMod Found: {HerobrineSoftDep.enabled}");
+            if (HerobrineSoftDep.enabled)
+            {
+                Harmony.PatchAll(typeof(HerobrineKillPatch));
+                Harmony.PatchAll(typeof(MinecraftPlayerExplodePatch));
+                //
+                HerobrineSoftDep.CoronerRegister();
             }
 
             PluginLogger.LogInfo($"UsualScrap Found: {UsualScrapSoftDep.enabled}");